Create Transaction AML
This code uses the CEL library to open an Excel document called "transactions.xlsx" and access its first worksheet. It then iterates through the rows of the sheet, reading the values of the 'borrower' and 'transaction amount' columns.
It stores the transactions for each borrower in a map, where the key is the borrower's name and the value is a vector of transaction amounts.

It then iterates through the borrowers and calculates the mean and standard deviation of their transactions using the formulas for mean and standard deviation. It then iterates through the transactions again and checks whether each transaction is an outlier using the formula for an outlier, which is any value that is more than 3 standard deviations away from the mean.
If the code finds an outlier, it will print the borrower name and the transaction amount.

Please note that the above code is an example and may not work as is. The column number and row number might be different in your case and you might have to use different library to read excel file.