Create between 8500 and 10000
In this version of the code, the function check_large_transactions takes 3 parameters, account_number, amounts and 2 threshold values lower_threshold and upper_threshold and it checks if there are more than 5 transactions that fall between the range of these two thresholds. If it finds such transactions, it prints the account number, the number of transactions found and the rows where these transactions were found.