Create base line +-<> Base Amount
In this version of the code, the function check_similar_transactions takes 3 parameters, account_number, amounts and base_amount and it checks if there are more than 5 transactions that are within $100 of the base amount. If it finds such transactions, it prints the account number, the number of transactions found and the rows where these transactions were found.
It is important to note that if you have a large data set, this implementation can be quite time-consuming as it will check for similar transactions for every transaction in the dataset.